+++ src/Bpp/Seq/Io/Maf/QualityFilterMafIterator.cpp
@@ -76,7 +76,7 @@ MafBlock* QualityFilterMafIterator::analyseCurrentBlock_() throw (Exception)
         size_t nr = aln.size();
         size_t nc = block->getNumberOfSites();
         //First we create a mask:
-        vector<unsigned int> pos;
+        vector<size_t> pos;
         vector<int> col(nr);
         //Reset window:
         window_.clear();
@@ -119,7 +119,7 @@ MafBlock* QualityFilterMafIterator::analyseCurrentBlock_() throw (Exception)
           }
       
           //Move forward:
-          for (unsigned int k = 0; k < step_; ++k) {
+          for (size_t k = 0; k < step_; ++k) {
             for (size_t j = 0; j < nr; ++j) {
               col[j] = aln[j][i];
             }  
@@ -182,7 +182,7 @@ MafBlock* QualityFilterMafIterator::analyseCurrentBlock_() throw (Exception)
               MafBlock* newBlock = new MafBlock();
               newBlock->setScore(block->getScore());
               newBlock->setPass(block->getPass());
-              for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+              for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
                 MafSequence* subseq;
                 if (i == 0) {
                   subseq = block->getSequence(j).subSequence(0, pos[i]);
@@ -199,7 +199,7 @@ MafBlock* QualityFilterMafIterator::analyseCurrentBlock_() throw (Exception)
               MafBlock* outBlock = new MafBlock();
               outBlock->setScore(block->getScore());
               outBlock->setPass(block->getPass());
-              for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+              for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
                 MafSequence* outseq = block->getSequence(j).subSequence(pos[i], pos[i + 1] - pos[i]);
                 outBlock->addSequence(*outseq);
                 delete outseq;
@@ -212,7 +212,7 @@ MafBlock* QualityFilterMafIterator::analyseCurrentBlock_() throw (Exception)
             MafBlock* newBlock = new MafBlock();
             newBlock->setScore(block->getScore());
             newBlock->setPass(block->getPass());
-            for (unsigned int j = 0; j < block->getNumberOfSequences(); ++j) {
+            for (size_t j = 0; j < block->getNumberOfSequences(); ++j) {
               MafSequence* subseq;
               subseq = block->getSequence(j).subSequence(pos[pos.size() - 1], block->getNumberOfSites() - pos[pos.size() - 1]);
               newBlock->addSequence(*subseq);
